+++ anim/data/counter.txt
