+++ anim/data/shaders/fullspectrumcyber.py
