+++ anim/data/template/doomfire.py
